chore(catalog-generator): Add tests for the Catalog generator
At this time, we're adding tests for the `ConfigBean` class and the
`GenerateCommandTest` class.

fix: https://github.com/KaotoIO/kaoto/issues/1250
